Add tests and betamax files.
