+++ tests/betamax/TestSingleUserCRUD-setUp.json
@@ -1,0 +1,59 @@
+{
+  "http_interactions": [
+    {
+      "recorded_at": "2017-04-17T14:10:06",
+      "request": {
+        "body": {
+          "encoding": "utf-8",
+          "string": ""
+        },
+        "headers": {
+          "Accept": "*/*",
+          "Accept-Encoding": "gzip, deflate",
+          "Authorization": "Basic ZmFjZXRvZUBmYWNldG9lLmNvbS5hdS90b2tlbjpKTGQ1WWVsb2hXR21BNmpmZmhyT2NneUlFSnpBYUdPQTBwWVNLSk5x",
+          "Connection": "keep-alive",
+          "Content-type": "application/json",
+          "User-Agent": "Zenpy/1.1"
+        },
+        "method": "GET",
+        "uri": "https://facetoehelp.zendesk.com/api/v2/search.json?query=type:%22user%22"
+      },
+      "response": {
+        "body": {
+          "base64_string": "H4sIAAAAAAAAA21TQZKbMBD8i862QZg1Xk6pHPKBbC5JpVRjGINiIVHSyMna5b9nZDDZ3coNtWZaPT3NVXgM0VAQ9Y+r0K2oC1nk1U7uqny/EtEbUYueaAx1lh2hQXLYoxk3F7QthtOmcUMGo87ORRYD+pC96d/8Cs6KlbAwINN8mdoZwAF0Ij46l2+qp7KS5T5/llIWVfnpAP4Al0TMlY1HIGwVEJcXuazWebmWxUu+q7dVncvvXBPH9j811Yvc1k9FXT6nGtIDqouzScdnsN3JnRgd+ztiozErEXrw/NIdUzYOB/SPK8bIPQ7GNWBQJa/kSkwnZkW7/vaVOZ3vwOoLkHZ2Ktrl5XZb7bd7NtQ7kyRAO+jkzBm9PmpkKvIR2Zg/hN6CuTdOsgi6tJyfKwFGA39OMDSkz0w19U3aRX0EE5hmHgU6tOzbDBoIpIzrtP3g5genfjvFiybnFUTqFVo4mKRwdkl3Fij6xbYWiZe56LKOcDk0MZAbVBr6zUSDa9EDP/BQT7o5ISlOInnNc3Fo5tecNa9q9PrMC1aciIEHYvZ5okdDyse7UUMMY8onq55Lmx5IJbYFSWlV7L1pmfB649XcfwRFr2PaULoWN3b9Hvp/4/GG1MivLdHweNYuhndg42IyXt7+AumVr05fAwAA",
+          "encoding": "UTF-8",
+          "string": ""
+        },
+        "headers": {
+          "Cache-Control": "must-revalidate, private, max-age=0",
+          "Connection": "keep-alive",
+          "Content-Encoding": "gzip",
+          "Content-Type": "application/json; charset=UTF-8",
+          "Date": "Mon, 17 Apr 2017 14:10:09 GMT",
+          "ETag": "W/\"d33155872de19118ba0f9a8f13f5cf45\"",
+          "Server": "nginx",
+          "Strict-Transport-Security": "max-age=31536000;",
+          "Transfer-Encoding": "chunked",
+          "X-Content-Type-Options": "nosniff",
+          "X-Frame-Options": "SAMEORIGIN",
+          "X-Rack-Cache": "miss",
+          "X-Rate-Limit": "400",
+          "X-Rate-Limit-Remaining": "397",
+          "X-Request-Id": "d6ce20dc-66f1-463f-c6b2-b8ca3a6bb1dd",
+          "X-Runtime": "0.254160",
+          "X-UA-Compatible": "IE=Edge,chrome=1",
+          "X-Zendesk-API-Version": "v2",
+          "X-Zendesk-Application-Version": "v19.29",
+          "X-Zendesk-Origin-Server": "app13.pod6.iad1.zdsys.com",
+          "X-Zendesk-Request-Id": "4f848172536858236782"
+        },
+        "status": {
+          "code": 200,
+          "message": "OK"
+        },
+        "url": "https://facetoehelp.zendesk.com/api/v2/search.json?query=type:%22user%22"
+      }
+    }
+  ],
+  "recorded_with": "betamax/0.8.0"
+}
