+++ tests/betamax/TestTicketApiExceptions.test_throws_on_invalid_type.json
@@ -1,0 +1,4 @@
+{
+  "http_interactions": [],
+  "recorded_with": "betamax/0.8.0"
+}
