+++ tests/betamax/TestTicketProperties-tearDown.json
@@ -1,0 +1,114 @@
+{
+  "http_interactions": [
+    {
+      "recorded_at": "2017-04-17T14:10:48",
+      "request": {
+        "body": {
+          "encoding": "utf-8",
+          "string": ""
+        },
+        "headers": {
+          "Accept": "*/*",
+          "Accept-Encoding": "gzip, deflate",
+          "Authorization": "Basic ZmFjZXRvZUBmYWNldG9lLmNvbS5hdS90b2tlbjpKTGQ1WWVsb2hXR21BNmpmZmhyT2NneUlFSnpBYUdPQTBwWVNLSk5x",
+          "Connection": "keep-alive",
+          "Content-type": "application/json",
+          "User-Agent": "Zenpy/1.1"
+        },
+        "method": "GET",
+        "uri": "https://facetoehelp.zendesk.com/api/v2/tickets.json?&include=users,groups,organizations"
+      },
+      "response": {
+        "body": {
+          "base64_string": "H4sIAAAAAAAAA61Vy3LbOBD8F5xl8yFaVHjayiE/kOwlqRQKJEcSIhDg4iHHdunf0+DL9NrlraT2xOIAM+gZdDeemJfNmbxj1bcnFqxiFTt537sqSQ6iIW/oRKq/fSTdkjvfNqZLRC+TS55Micluf3f7wxnNNky2rMLvhtFPT1YLxWNEB6U27CIFq55YcxJaUzwGZZDiTLANxZWDNR2+1w3zZvzauC9mXxFsLAlPLRceuXmalTdpcZOVX7KiytKqKL+iWOjb/9zjH3ocN2Jyof5BTSzoyfmb3pqerJfkUMuKe/7eOsbRWNl7ic6Rf5L6GNN6K42V/mE5wgsfMF2G0nFElhrZS9I4dMRg6Z+Aw8kOs8qzPC132a5M95hNqDvp31wSzsmjJnqdZOxRaPkoIrBhNdulxXZb7rd7lDxaE/ohXCCeDbHGKCVqY4U3EUSkwvcNOxgbOu5NL5vVLWJEtaJuFTkJx6VuZIuekHoQyhGY4HgfaiUbVnkbEGgDDVc3Nu0FhjUc0wTnTccPktR8sgN2B/INHQAVJrtM8yQsfrk4WqIOJ67wrirUVuh2wFhmeZblaFIoZe75xL5aNOcJ6hUYgiM7CuDVFfyWIoY6yeoKZ11o0YFz7NMoKLCAOiGjBg7GpLflXVFmxT79kAFqWfxVC1uLxyg17Hyb9vmXdFdtyyrN3qX9trrLq+JD3ONlR/zR6Ijjo9DHszlHtp6GyKQGDBcCG2Jch64mOy8hFkU57lOmEWpkXrZh4x+qkr75+zNqvk9Aa1SEINpORjlcyErcPW5qpMkbxvHMFaGkwD2NKCI/Lig15rkB+0K/8Rc0GYQ2cVIJ57kyR6lfm8h6UveGR/ZBDCL4EyctQHnUnqYE3UHSdhlbSx6XueDSBl4yb57YHZteSaYzLQ1qm9GPVsotfMDKgfdzAaPVA4enXOBrHIyInH9W2ZwQbfFFqy442E27mgiY73mstowjsnXR3dM1CmFwh1EJv0X8MS+ZLWX9HDzbzCSCz6HvjfW4+5YUwdEXQH9O9UUO+VcW+1hT8A/aeZGerPxz3dgLW516wyvZP6CziX7L0zoRcAq/usY3SP//zAIzNnAazSO+yXDfo+vz6/Bvm34h6tmsR9JowOc9+LeYhaWLNAEvwCrYmBClmF1/ASW0/71yCAAA",
+          "encoding": "UTF-8",
+          "string": ""
+        },
+        "headers": {
+          "Cache-Control": "must-revalidate, private, max-age=0",
+          "Connection": "keep-alive",
+          "Content-Encoding": "gzip",
+          "Content-Type": "application/json; charset=UTF-8",
+          "Date": "Mon, 17 Apr 2017 14:10:52 GMT",
+          "ETag": "W/\"19c6843419439b0c23ae5f6363c97e60\"",
+          "Server": "nginx",
+          "Strict-Transport-Security": "max-age=31536000;",
+          "Transfer-Encoding": "chunked",
+          "X-Content-Type-Options": "nosniff",
+          "X-Frame-Options": "SAMEORIGIN",
+          "X-Rack-Cache": "miss",
+          "X-Rate-Limit": "400",
+          "X-Rate-Limit-Remaining": "364",
+          "X-Request-Id": "40a42442-bec8-4670-c6d3-b8ca3a6b2498",
+          "X-Runtime": "0.249166",
+          "X-UA-Compatible": "IE=Edge,chrome=1",
+          "X-Zendesk-API-Version": "v2",
+          "X-Zendesk-Application-Version": "v19.29",
+          "X-Zendesk-Origin-Server": "app19.pod6.iad1.zdsys.com",
+          "X-Zendesk-Request-Id": "9c024a882acec858fef9"
+        },
+        "status": {
+          "code": 200,
+          "message": "OK"
+        },
+        "url": "https://facetoehelp.zendesk.com/api/v2/tickets.json?&include=users,groups,organizations"
+      }
+    },
+    {
+      "recorded_at": "2017-04-17T14:10:49",
+      "request": {
+        "body": {
+          "encoding": "utf-8",
+          "string": ""
+        },
+        "headers": {
+          "Accept": "*/*",
+          "Accept-Encoding": "gzip, deflate",
+          "Authorization": "Basic ZmFjZXRvZUBmYWNldG9lLmNvbS5hdS90b2tlbjpKTGQ1WWVsb2hXR21BNmpmZmhyT2NneUlFSnpBYUdPQTBwWVNLSk5x",
+          "Connection": "keep-alive",
+          "Content-Length": "0",
+          "Content-type": "application/json",
+          "User-Agent": "Zenpy/1.1"
+        },
+        "method": "DELETE",
+        "uri": "https://facetoehelp.zendesk.com/api/v2/tickets/destroy_many.json?ids=685&include=users,groups,organizations"
+      },
+      "response": {
+        "body": {
+          "base64_string": "H4sIAAAAAAAAA4WOzQ6CMBCE32XPhBIUqryMKe3yZ6GV3XqQ8O5uosSjx8l8mW82mEJ7IzacCJoNRgcNlLrWxrbni8VK22txwg5b57q6wFNVOg0ZpNULODBHapTqjEUOOKCP+QsXh3TPbZiViaN6luqnQFL/xvOJwiIGDmzEsSTvM4hr6FckufjJx2F4JEzoBJ+lNT0egMDJ85ff9zdJWjh/5wAAAA==",
+          "encoding": "UTF-8",
+          "string": ""
+        },
+        "headers": {
+          "Cache-Control": "max-age=0, private, must-revalidate",
+          "Connection": "keep-alive",
+          "Content-Encoding": "gzip",
+          "Content-Type": "application/json; charset=UTF-8",
+          "Date": "Mon, 17 Apr 2017 14:10:52 GMT",
+          "ETag": "W/\"12661c21e5482ebe6dbbc5a4d2a470b1\"",
+          "Server": "nginx",
+          "Strict-Transport-Security": "max-age=31536000;",
+          "Transfer-Encoding": "chunked",
+          "X-Content-Type-Options": "nosniff",
+          "X-Frame-Options": "SAMEORIGIN",
+          "X-Rack-Cache": "invalidate, pass",
+          "X-Rate-Limit": "400",
+          "X-Rate-Limit-Remaining": "363",
+          "X-Request-Id": "36ed9f61-2630-4917-c9db-b8ca3a6b2498",
+          "X-Runtime": "0.215087",
+          "X-UA-Compatible": "IE=Edge,chrome=1",
+          "X-Zendesk-API-Version": "v2",
+          "X-Zendesk-Application-Version": "v19.29",
+          "X-Zendesk-Origin-Server": "app9.pod6.iad1.zdsys.com",
+          "X-Zendesk-Request-Id": "0bfccd0f721318e64bce"
+        },
+        "status": {
+          "code": 200,
+          "message": "OK"
+        },
+        "url": "https://facetoehelp.zendesk.com/api/v2/tickets/destroy_many.json?ids=685&include=users,groups,organizations"
+      }
+    }
+  ],
+  "recorded_with": "betamax/0.8.0"
+}
